refactor: split native npm bindings
