+++ crates/vize_vitrine/src/napi/sfc.rs
@@ -1,8 +1,5 @@
 //! NAPI bindings for SFC parsing and compilation.
 //!
-//! Provides parseSfc, compileSfc, compileSfcBatch, and
-//! compileSfcBatchWithResults functions for Vue Single File Components.
-//!
 //! FFI boundary code: uses std types for JavaScript interop.
 #![allow(
     clippy::disallowed_types,
@@ -10,1081 +7,18 @@
     clippy::disallowed_macros
 )]
 
-use glob::glob;
-use napi::bindgen_prelude::{Error, Result, Status};
-use napi_derive::napi;
-use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
-use serde_json::{Value, json};
-use std::{
-    fs,
-    sync::atomic::{AtomicUsize, Ordering},
-};
-use vize_carton::cstr;
-
-#[napi(object)]
-pub struct StyleBlockNapi {
-    pub content: String,
-    pub src: Option<String>,
-    pub lang: Option<String>,
-    pub scoped: bool,
-    pub module: bool,
-    pub module_name: Option<String>,
-    pub index: u32,
-}
-
-#[napi(object)]
-pub struct SfcBlockAttributeNapi {
-    pub name: String,
-    pub value: Option<String>,
-}
-
-#[napi(object)]
-pub struct CustomBlockNapi {
-    pub block_type: String,
-    pub content: String,
-    pub src: Option<String>,
-    pub attrs: Vec<SfcBlockAttributeNapi>,
-    pub index: u32,
-}
-
-#[napi(object)]
-pub struct SfcSrcInfoNapi {
-    pub script_src: Option<String>,
-    pub template_src: Option<String>,
-}
-
-#[napi(object)]
-pub struct TemplateAssetUrlNapi {
-    pub url: String,
-    pub var_name: String,
-}
-
-#[napi(object)]
-pub struct TemplateAssetTagRuleNapi {
-    pub tag: String,
-    pub attrs: Vec<String>,
-}
-
-#[napi(object)]
-pub struct MacroArtifactNapi {
-    pub kind: String,
-    pub name: String,
-    pub source: String,
-    pub content: String,
-    pub module_code: Option<String>,
-    pub start: u32,
-    pub end: u32,
-}
-
-fn macro_artifacts_to_napi(
-    artifacts: Vec<vize_atelier_sfc::SfcMacroArtifact>,
-) -> Vec<MacroArtifactNapi> {
-    artifacts
-        .into_iter()
-        .map(|artifact| MacroArtifactNapi {
-            kind: artifact.kind.into(),
-            name: artifact.name.into(),
-            source: artifact.source.into(),
-            content: artifact.content.into(),
-            module_code: artifact.module_code.map(Into::into),
-            start: artifact.start as u32,
-            end: artifact.end as u32,
-        })
-        .collect()
-}
-
-fn style_blocks_to_napi(styles: &[vize_atelier_sfc::SfcStyleBlock]) -> Vec<StyleBlockNapi> {
-    styles
-        .iter()
-        .enumerate()
-        .map(|(index, style)| {
-            let module_attr = style.attrs.get("module");
-            let module_name = module_attr.and_then(|value| {
-                let value = value.as_ref();
-                if value.is_empty() {
-                    None
-                } else {
-                    Some(value.into())
-                }
-            });
-
-            StyleBlockNapi {
-                content: style.content.as_ref().into(),
-                src: style.src.as_deref().map(Into::into),
-                lang: style.lang.as_deref().map(Into::into),
-                scoped: style.scoped,
-                module: module_attr.is_some(),
-                module_name,
-                index: index as u32,
-            }
-        })
-        .collect()
-}
-
-fn custom_blocks_to_napi(blocks: &[vize_atelier_sfc::SfcCustomBlock]) -> Vec<CustomBlockNapi> {
-    blocks
-        .iter()
-        .enumerate()
-        .map(|(index, block)| {
-            let mut attrs = block
-                .attrs
-                .iter()
-                .map(|(name, value)| SfcBlockAttributeNapi {
-                    name: name.as_ref().into(),
-                    value: (!value.is_empty()).then(|| value.as_ref().into()),
-                })
-                .collect::<Vec<_>>();
-            attrs.sort_by(|left, right| left.name.cmp(&right.name));
-            CustomBlockNapi {
-                block_type: block.block_type.as_ref().into(),
-                content: block.content.as_ref().into(),
-                src: block.attrs.get("src").map(|value| value.as_ref().into()),
-                attrs,
-                index: index as u32,
-            }
-        })
-        .collect()
-}
-
-impl From<vize_atelier_sfc::BundlerStyleBlock> for StyleBlockNapi {
-    fn from(block: vize_atelier_sfc::BundlerStyleBlock) -> Self {
-        Self {
-            content: block.content.into(),
-            src: block.src.map(Into::into),
-            lang: block.lang.map(Into::into),
-            scoped: block.scoped,
-            module: block.module,
-            module_name: block.module_name.map(Into::into),
-            index: block.index,
-        }
-    }
-}
-
-impl From<vize_atelier_sfc::SfcBlockAttribute> for SfcBlockAttributeNapi {
-    fn from(attr: vize_atelier_sfc::SfcBlockAttribute) -> Self {
-        Self {
-            name: attr.name.into(),
-            value: attr.value.map(Into::into),
-        }
-    }
-}
-
-impl From<vize_atelier_sfc::BundlerCustomBlock> for CustomBlockNapi {
-    fn from(block: vize_atelier_sfc::BundlerCustomBlock) -> Self {
-        Self {
-            block_type: block.block_type.into(),
-            content: block.content.into(),
-            src: block.src.map(Into::into),
-            attrs: block.attrs.into_iter().map(Into::into).collect(),
-            index: block.index,
-        }
-    }
-}
-
-impl From<vize_atelier_sfc::SfcSrcInfo> for SfcSrcInfoNapi {
-    fn from(info: vize_atelier_sfc::SfcSrcInfo) -> Self {
-        Self {
-            script_src: info.script_src.map(Into::into),
-            template_src: info.template_src.map(Into::into),
-        }
-    }
-}
-
-impl From<vize_atelier_sfc::TemplateAssetUrl> for TemplateAssetUrlNapi {
-    fn from(url: vize_atelier_sfc::TemplateAssetUrl) -> Self {
-        Self {
-            url: url.url.into(),
-            var_name: url.var_name.into(),
-        }
-    }
-}
-
-impl From<TemplateAssetTagRuleNapi> for vize_atelier_sfc::TemplateAssetTagRule {
-    fn from(rule: TemplateAssetTagRuleNapi) -> Self {
-        Self {
-            tag: rule.tag.into(),
-            attrs: rule.attrs.into_iter().map(Into::into).collect(),
-        }
-    }
-}
-
-/// SFC parse options for NAPI
-#[napi(object)]
-#[derive(Default)]
-pub struct SfcParseOptionsNapi {
-    pub filename: Option<String>,
-}
-
-/// SFC compile options for NAPI
-#[napi(object)]
-#[derive(Default)]
-pub struct SfcCompileOptionsNapi {
-    pub filename: Option<String>,
-    pub source_map: Option<bool>,
-    pub ssr: Option<bool>,
-    pub vapor: Option<bool>,
-    pub custom_renderer: Option<bool>,
-    /// Preserve TypeScript in output when true
-    pub is_ts: Option<bool>,
-    /// Scope ID for scoped CSS (e.g., "data-v-abc123")
-    pub scope_id: Option<String>,
-}
-
-/// SFC compile result for NAPI
-#[napi(object)]
-pub struct SfcCompileResultNapi {
-    /// Generated JavaScript code
-    pub code: String,
-    /// Generated CSS (if any)
-    pub css: Option<String>,
-    /// Compilation errors
-    pub errors: Vec<String>,
-    /// Compilation warnings
-    pub warnings: Vec<String>,
-    /// Hash of template content (for HMR)
-    pub template_hash: Option<String>,
-    /// Hash of style content (for HMR)
-    pub style_hash: Option<String>,
-    /// Hash of script content (for HMR)
-    pub script_hash: Option<String>,
-    /// Whether the file has scoped styles
-    pub has_scoped: bool,
-    /// Per-block style metadata
-    pub styles: Vec<StyleBlockNapi>,
-    /// Custom block metadata
-    pub custom_blocks: Vec<CustomBlockNapi>,
-    /// Compile-time macro artifacts
-    pub macro_artifacts: Vec<MacroArtifactNapi>,
-}
-
-/// Batch compile options for NAPI
-#[napi(object)]
-#[derive(Default)]
-pub struct BatchCompileOptionsNapi {
-    pub ssr: Option<bool>,
-    pub vapor: Option<bool>,
-    pub custom_renderer: Option<bool>,
-    /// Preserve TypeScript in output when true
-    pub is_ts: Option<bool>,
-    pub threads: Option<u32>,
-}
-
-/// Batch compile result for NAPI
-#[napi(object)]
-pub struct BatchCompileResultNapi {
-    /// Number of files compiled successfully
-    pub success: u32,
-    /// Number of files that failed
-    pub failed: u32,
-    /// Total input bytes
-    pub input_bytes: u32,
-    /// Total output bytes
-    pub output_bytes: u32,
-    /// Compilation time in milliseconds
-    pub time_ms: f64,
-}
-
-/// Input file for batch compilation with results
-#[napi(object)]
-pub struct BatchFileInputNapi {
-    /// File path
-    pub path: String,
-    /// Source code
-    pub source: String,
-}
-
-/// Per-file result from batch compilation
-#[napi(object)]
-pub struct BatchFileResultNapi {
-    /// File path
-    pub path: String,
-    /// Generated JavaScript code
-    pub code: String,
-    /// Generated CSS (if any)
-    pub css: Option<String>,
-    /// Scope ID for scoped styles
-    pub scope_id: String,
-    /// Whether the file has scoped styles
-    pub has_scoped: bool,
-    /// Compilation errors
-    pub errors: Vec<String>,
-    /// Compilation warnings
-    pub warnings: Vec<String>,
-    /// Hash of template content (for HMR)
-    pub template_hash: Option<String>,
-    /// Hash of style content (for HMR)
-    pub style_hash: Option<String>,
-    /// Hash of script content (for HMR)
-    pub script_hash: Option<String>,
-    /// Per-block style metadata
-    pub styles: Vec<StyleBlockNapi>,
-    /// Custom block metadata
-    pub custom_blocks: Vec<CustomBlockNapi>,
-    /// Compile-time macro artifacts
-    pub macro_artifacts: Vec<MacroArtifactNapi>,
-}
-
-/// Batch compile result with per-file results
-#[napi(object)]
-pub struct BatchCompileResultWithFilesNapi {
-    /// Per-file compilation results
-    pub results: Vec<BatchFileResultNapi>,
-    /// Number of files compiled successfully
-    pub success_count: u32,
-    /// Number of files that failed
-    pub failed_count: u32,
-    /// Compilation time in milliseconds
-    pub time_ms: f64,
-}
-
-/// Parse SFC (.vue file) - returns lightweight result for speed
-#[napi(js_name = "parseSfc")]
-pub fn parse_sfc(source: String, options: Option<SfcParseOptionsNapi>) -> Result<Value> {
-    use vize_atelier_sfc::{SfcParseOptions, parse_sfc as sfc_parse};
-
-    let opts = options.unwrap_or_default();
-    let parse_opts = SfcParseOptions {
-        filename: opts
-            .filename
-            .unwrap_or_else(|| "anonymous.vue".to_string())
-            .into(),
-        ..Default::default()
-    };
-
-    match sfc_parse(&source, parse_opts) {
-        Ok(descriptor) => {
-            let template = if let Some(ref template) = descriptor.template {
-                json!({
-                    "content": template.content.as_ref(),
-                    "lang": template.lang.as_deref(),
-                })
-            } else {
-                Value::Null
-            };
-
-            let script = if let Some(ref script) = descriptor.script {
-                json!({
-                    "content": script.content.as_ref(),
-                    "lang": script.lang.as_deref(),
-                    "setup": script.setup,
-                })
-            } else {
-                Value::Null
-            };
-
-            let script_setup = if let Some(ref script_setup) = descriptor.script_setup {
-                json!({
-                    "content": script_setup.content.as_ref(),
-                    "lang": script_setup.lang.as_deref(),
-                    "setup": script_setup.setup,
-                })
-            } else {
-                Value::Null
-            };
-
-            let styles = descriptor
-                .styles
-                .iter()
-                .map(|style| {
-                    json!({
-                        "content": style.content.as_ref(),
-                        "lang": style.lang.as_deref(),
-                        "scoped": style.scoped,
-                        "module": style.module.as_deref(),
-                    })
-                })
-                .collect::<Vec<_>>();
-            let custom_blocks = descriptor
-                .custom_blocks
-                .iter()
-                .map(|block| {
-                    json!({
-                        "type": block.block_type.as_ref(),
-                        "content": block.content.as_ref(),
-                    })
-                })
-                .collect::<Vec<_>>();
-
-            Ok(json!({
-                "filename": descriptor.filename.as_ref(),
-                "source": descriptor.source.as_ref(),
-                "template": template,
-                "script": script,
-                "scriptSetup": script_setup,
-                "styles": styles,
-                "customBlocks": custom_blocks,
-            }))
-        }
-        Err(e) => Err(Error::new(Status::GenericFailure, e.message.to_string())),
-    }
-}
-
-#[napi(js_name = "generateSfcScopeId")]
-pub fn generate_sfc_scope_id(
-    filename: String,
-    root: Option<String>,
-    is_production: Option<bool>,
-    source: Option<String>,
-) -> String {
-    vize_atelier_sfc::generate_bundler_scope_id(
-        &filename,
-        root.as_deref(),
-        is_production.unwrap_or(false),
-        source.as_deref(),
-    )
-    .into()
-}
-
-#[napi(js_name = "extractSfcStyleBlocks")]
-pub fn extract_sfc_style_blocks(source: String, filename: Option<String>) -> Vec<StyleBlockNapi> {
-    vize_atelier_sfc::extract_style_blocks(&source, filename.as_deref())
-        .into_iter()
-        .map(Into::into)
-        .collect()
-}
-
-#[napi(js_name = "extractSfcCustomBlocks")]
-pub fn extract_sfc_custom_blocks(source: String, filename: Option<String>) -> Vec<CustomBlockNapi> {
-    vize_atelier_sfc::extract_custom_blocks(&source, filename.as_deref())
-        .into_iter()
-        .map(Into::into)
-        .collect()
-}
-
-#[napi(js_name = "extractSfcSrcInfo")]
-pub fn extract_sfc_src_info(source: String, filename: Option<String>) -> SfcSrcInfoNapi {
-    vize_atelier_sfc::extract_src_info(&source, filename.as_deref()).into()
-}
-
-#[napi(js_name = "hasSfcScopedStyle")]
-pub fn has_sfc_scoped_style(source: String, filename: Option<String>) -> bool {
-    vize_atelier_sfc::has_scoped_style(&source, filename.as_deref())
-}
-
-#[napi(js_name = "isSfcImportableAssetUrl")]
-pub fn is_sfc_importable_asset_url(url: String) -> bool {
-    vize_atelier_sfc::is_importable_asset_url(&url)
-}
-
-#[napi(js_name = "collectSfcTemplateAssetUrls")]
-pub fn collect_sfc_template_asset_urls(
-    source: String,
-    rules: Option<Vec<TemplateAssetTagRuleNapi>>,
-    filename: Option<String>,
-) -> Vec<TemplateAssetUrlNapi> {
-    let rules = rules.map(|rules| rules.into_iter().map(Into::into).collect::<Vec<_>>());
-    vize_atelier_sfc::collect_template_asset_urls(&source, rules.as_deref(), filename.as_deref())
-        .into_iter()
-        .map(Into::into)
-        .collect()
-}
-
-#[napi(js_name = "stripSfcScopedCssComments")]
-pub fn strip_sfc_scoped_css_comments(css: String) -> String {
-    vize_atelier_sfc::strip_css_comments_for_scoped(&css).into()
-}
-
-#[napi(js_name = "wrapSfcScopedPreprocessorStyle")]
-pub fn wrap_sfc_scoped_preprocessor_style(
-    content: String,
-    scoped: Option<String>,
-    lang: Option<String>,
-) -> String {
-    vize_atelier_sfc::wrap_scoped_preprocessor_style(&content, scoped.as_deref(), lang.as_deref())
-        .into()
-}
-
-/// Compile SFC (.vue file) to JavaScript - main use case
-#[napi(js_name = "compileSfc")]
-pub fn compile_sfc(
-    source: String,
-    options: Option<SfcCompileOptionsNapi>,
-) -> Result<SfcCompileResultNapi> {
-    use vize_atelier_sfc::{
-        ScriptCompileOptions, SfcCompileOptions, SfcParseOptions, StyleCompileOptions,
-        TemplateCompileOptions, compile_sfc as sfc_compile, parse_sfc as sfc_parse,
-    };
-
-    let opts = options.unwrap_or_default();
-    let filename: vize_carton::CompactString = opts
-        .filename
-        .unwrap_or_else(|| "anonymous.vue".to_string())
-        .into();
-
-    // Parse
-    let parse_opts = SfcParseOptions {
-        filename: filename.clone(),
-        ..Default::default()
-    };
-
-    let descriptor = match sfc_parse(&source, parse_opts) {
-        Ok(d) => d,
-        Err(e) => {
-            return Ok(SfcCompileResultNapi {
-                code: String::new(),
-                css: None,
-                errors: vec![e.message.into()],
-                warnings: vec![],
-                template_hash: None,
-                style_hash: None,
-                script_hash: None,
-                has_scoped: false,
-                styles: vec![],
-                custom_blocks: vec![],
-                macro_artifacts: vec![],
-            });
-        }
-    };
-
-    let template_hash: Option<String> = descriptor.template_hash().map(Into::into);
-    let style_hash: Option<String> = descriptor.style_hash().map(Into::into);
-    let script_hash: Option<String> = descriptor.script_hash().map(Into::into);
-    let styles = style_blocks_to_napi(&descriptor.styles);
-    let custom_blocks = custom_blocks_to_napi(&descriptor.custom_blocks);
-
-    // Compile
-    let has_scoped = descriptor.styles.iter().any(|s| s.scoped);
-    let vapor = opts.vapor.unwrap_or(false);
-    let is_ts = opts.is_ts.unwrap_or(false);
-
-    // Extract scope_id from options (strip "data-v-" prefix if present)
-    let external_scope_id: Option<vize_carton::CompactString> = opts
-        .scope_id
-        .as_ref()
-        .map(|sid| sid.strip_prefix("data-v-").unwrap_or(sid).into());
-
-    // Create compiler options with scope_id for scoped CSS
-    let template_compiler_options = if has_scoped {
-        external_scope_id
-            .as_ref()
-            .map(|scope_id| vize_atelier_dom::DomCompilerOptions {
-                scope_id: Some(cstr!("data-v-{scope_id}")),
-                ..Default::default()
-            })
-    } else {
-        None
-    };
-
-    let compile_opts = SfcCompileOptions {
-        parse: SfcParseOptions {
-            filename: filename.clone(),
-            ..Default::default()
-        },
-        script: ScriptCompileOptions {
-            id: Some(filename.clone()),
-            is_ts,
-            ..Default::default()
-        },
-        template: TemplateCompileOptions {
-            id: Some(filename.clone()),
-            scoped: has_scoped,
-            ssr: opts.ssr.unwrap_or(false),
-            is_ts,
-            custom_renderer: opts.custom_renderer.unwrap_or(false),
-            compiler_options: template_compiler_options,
-            ..Default::default()
-        },
-        style: StyleCompileOptions {
-            id: filename,
-            scoped: has_scoped,
-            ..Default::default()
-        },
-        vapor,
-        scope_id: external_scope_id,
-    };
-
-    match sfc_compile(&descriptor, compile_opts) {
-        Ok(result) => {
-            let macro_artifacts = macro_artifacts_to_napi(result.macro_artifacts);
-            Ok(SfcCompileResultNapi {
-                code: result.code.into(),
-                css: result.css.map(Into::into),
-                errors: result
-                    .errors
-                    .into_iter()
-                    .map(|e| e.message.into())
-                    .collect(),
-                warnings: result
-                    .warnings
-                    .into_iter()
-                    .map(|e| e.message.into())
-                    .collect(),
-                template_hash: template_hash.clone(),
-                style_hash: style_hash.clone(),
-                script_hash: script_hash.clone(),
-                has_scoped,
-                styles,
-                custom_blocks,
-                macro_artifacts,
-            })
-        }
-        Err(e) => Ok(SfcCompileResultNapi {
-            code: String::new(),
-            css: None,
-            errors: vec![e.message.into()],
-            warnings: vec![],
-            template_hash,
-            style_hash,
-            script_hash,
-            has_scoped,
-            styles,
-            custom_blocks,
-            macro_artifacts: vec![],
-        }),
-    }
-}
-
-/// Batch compile SFC files matching a glob pattern (native multithreading)
-#[napi(js_name = "compileSfcBatch")]
-pub fn compile_sfc_batch(
-    pattern: String,
-    options: Option<BatchCompileOptionsNapi>,
-) -> Result<BatchCompileResultNapi> {
-    use std::time::Instant;
-    use vize_atelier_sfc::{
-        ScriptCompileOptions, SfcCompileOptions, SfcParseOptions, StyleCompileOptions,
-        TemplateCompileOptions, compile_sfc as sfc_compile, parse_sfc as sfc_parse,
-    };
-
-    let opts = options.unwrap_or_default();
-    let ssr = opts.ssr.unwrap_or(false);
-    let vapor = opts.vapor.unwrap_or(false);
-    let is_ts = opts.is_ts.unwrap_or(false);
-
-    // Configure thread pool if specified
-    if let Some(threads) = opts.threads {
-        rayon::ThreadPoolBuilder::new()
-            .num_threads(threads as usize)
-            .build_global()
-            .ok(); // Ignore if already configured
-    }
-
-    // Collect files matching the pattern
-    let files: Vec<_> = glob(&pattern)
-        .map_err(|e| {
-            Error::new(
-                Status::GenericFailure,
-                format!("Invalid glob pattern: {}", e),
-            )
-        })?
-        .filter_map(|entry| entry.ok())
-        .filter(|path| path.extension().is_some_and(|ext| ext == "vue"))
-        .collect();
-
-    if files.is_empty() {
-        return Err(Error::new(
-            Status::GenericFailure,
-            "No .vue files found matching the pattern",
-        ));
-    }
-
-    #[derive(Default)]
-    struct BatchStats {
-        success: usize,
-        failed: usize,
-        input_bytes: usize,
-        output_bytes: usize,
-    }
-
-    impl BatchStats {
-        fn failed() -> Self {
-            Self {
-                failed: 1,
-                ..Default::default()
-            }
-        }
-
-        fn failed_with_input(input_bytes: usize) -> Self {
-            Self {
-                failed: 1,
-                input_bytes,
-                ..Default::default()
-            }
-        }
-
-        fn success(input_bytes: usize, output_bytes: usize) -> Self {
-            Self {
-                success: 1,
-                input_bytes,
-                output_bytes,
-                failed: 0,
-            }
-        }
-
-        fn add(mut self, other: Self) -> Self {
-            self.success += other.success;
-            self.failed += other.failed;
-            self.input_bytes += other.input_bytes;
-            self.output_bytes += other.output_bytes;
-            self
-        }
-    }
-
-    let start = Instant::now();
-
-    // Compile files in parallel using rayon
-    let stats = files
-        .par_iter()
-        .map(|path| {
-            let source = match fs::read_to_string(path) {
-                Ok(s) => s,
-                Err(_) => {
-                    return BatchStats::failed();
-                }
-            };
-
-            let source_len = source.len();
-
-            let filename: vize_carton::CompactString = path.to_string_lossy().as_ref().into();
-
-            // Parse
-            let parse_opts = SfcParseOptions {
-                filename: filename.clone(),
-                ..Default::default()
-            };
-
-            let descriptor = match sfc_parse(&source, parse_opts) {
-                Ok(d) => d,
-                Err(_) => {
-                    return BatchStats::failed_with_input(source_len);
-                }
-            };
-
-            // Compile
-            let has_scoped = descriptor.styles.iter().any(|s| s.scoped);
-            let compile_opts = SfcCompileOptions {
-                parse: SfcParseOptions {
-                    filename: filename.clone(),
-                    ..Default::default()
-                },
-                script: ScriptCompileOptions {
-                    id: Some(filename.clone()),
-                    is_ts,
-                    ..Default::default()
-                },
-                template: TemplateCompileOptions {
-                    id: Some(filename.clone()),
-                    scoped: has_scoped,
-                    ssr,
-                    is_ts,
-                    ..Default::default()
-                },
-                style: StyleCompileOptions {
-                    id: filename,
-                    scoped: has_scoped,
-                    ..Default::default()
-                },
-                vapor,
-                scope_id: None,
-            };
-
-            match sfc_compile(&descriptor, compile_opts) {
-                Ok(result) => BatchStats::success(source_len, result.code.len()),
-                Err(_) => BatchStats::failed_with_input(source_len),
-            }
-        })
-        .reduce(BatchStats::default, BatchStats::add);
-
-    let elapsed = start.elapsed();
-
-    Ok(BatchCompileResultNapi {
-        success: stats.success as u32,
-        failed: stats.failed as u32,
-        input_bytes: stats.input_bytes as u32,
-        output_bytes: stats.output_bytes as u32,
-        time_ms: elapsed.as_secs_f64() * 1000.0,
-    })
-}
-
-/// Batch compile SFC files with per-file results (in-memory, native multithreading)
-#[napi(js_name = "compileSfcBatchWithResults")]
-pub fn compile_sfc_batch_with_results(
-    files: Vec<BatchFileInputNapi>,
-    options: Option<BatchCompileOptionsNapi>,
-) -> Result<BatchCompileResultWithFilesNapi> {
-    use std::sync::Mutex;
-    use std::time::Instant;
-    use vize_atelier_sfc::{
-        ScriptCompileOptions, SfcCompileOptions, SfcParseOptions, StyleCompileOptions,
-        TemplateCompileOptions, compile_sfc as sfc_compile, parse_sfc as sfc_parse,
-    };
-
-    let opts = options.unwrap_or_default();
-    let ssr = opts.ssr.unwrap_or(false);
-    let vapor = opts.vapor.unwrap_or(false);
-    let is_ts = opts.is_ts.unwrap_or(false);
-
-    // Configure thread pool if specified
-    if let Some(threads) = opts.threads {
-        rayon::ThreadPoolBuilder::new()
-            .num_threads(threads as usize)
-            .build_global()
-            .ok(); // Ignore if already configured
-    }
-
-    let results: Mutex<Vec<BatchFileResultNapi>> = Mutex::new(Vec::with_capacity(files.len()));
-    let success_count = AtomicUsize::new(0);
-    let failed_count = AtomicUsize::new(0);
-
-    let start = Instant::now();
-
-    let push_result = |result: BatchFileResultNapi| {
-        let mut guard = match results.lock() {
-            Ok(guard) => guard,
-            Err(poisoned) => poisoned.into_inner(),
-        };
-        guard.push(result);
-    };
-
-    // Compile files in parallel using rayon
-    files.par_iter().for_each(|file| {
-        let filename = &file.path;
-        let source = &file.source;
-
-        // Generate scope ID from filename using SHA-256 (matching JS-side generateScopeId)
-        let scope_id = {
-            use sha2::{Digest, Sha256};
-            let hash = Sha256::digest(filename.as_bytes());
-            // Take first 8 hex chars of the SHA-256 hash (same as JS: hash.slice(0, 8))
-            cstr!(
-                "{:02x}{:02x}{:02x}{:02x}",
-                hash[0],
-                hash[1],
-                hash[2],
-                hash[3]
-            )
-        };
-
-        let filename_cs: vize_carton::CompactString = filename.clone().into();
-
-        // Parse
-        let parse_opts = SfcParseOptions {
-            filename: filename_cs.clone(),
-            ..Default::default()
-        };
-
-        let descriptor = match sfc_parse(source, parse_opts) {
-            Ok(d) => d,
-            Err(e) => {
-                failed_count.fetch_add(1, Ordering::Relaxed);
-                push_result(BatchFileResultNapi {
-                    path: filename.clone(),
-                    code: String::new(),
-                    css: None,
-                    scope_id: scope_id.clone().into(),
-                    has_scoped: false,
-                    errors: vec![e.message.into()],
-                    warnings: vec![],
-                    template_hash: None,
-                    style_hash: None,
-                    script_hash: None,
-                    styles: vec![],
-                    custom_blocks: vec![],
-                    macro_artifacts: vec![],
-                });
-                return;
-            }
-        };
-
-        // Compute hashes for HMR
-        let template_hash: Option<String> = descriptor.template_hash().map(Into::into);
-        let style_hash: Option<String> = descriptor.style_hash().map(Into::into);
-        let script_hash: Option<String> = descriptor.script_hash().map(Into::into);
-        let styles = style_blocks_to_napi(&descriptor.styles);
-        let custom_blocks = custom_blocks_to_napi(&descriptor.custom_blocks);
-
-        // Compile
-        // Preserve TypeScript in output - let Vite/esbuild handle TS transformation
-        let actual_has_scoped = descriptor.styles.iter().any(|s| s.scoped);
-        // Create compiler options with scope_id for scoped CSS
-        let template_compiler_options = if actual_has_scoped {
-            Some(vize_atelier_dom::DomCompilerOptions {
-                scope_id: Some(cstr!("data-v-{scope_id}")),
-                ..Default::default()
-            })
-        } else {
-            None
-        };
-
-        let compile_opts = SfcCompileOptions {
-            parse: SfcParseOptions {
-                filename: filename_cs.clone(),
-                ..Default::default()
-            },
-            script: ScriptCompileOptions {
-                id: Some(filename_cs.clone()),
-                is_ts,
-                ..Default::default()
-            },
-            template: TemplateCompileOptions {
-                id: Some(filename_cs.clone()),
-                scoped: actual_has_scoped,
-                ssr,
-                is_ts,
-                custom_renderer: opts.custom_renderer.unwrap_or(false),
-                compiler_options: template_compiler_options,
-                ..Default::default()
-            },
-            style: StyleCompileOptions {
-                id: filename_cs,
-                scoped: actual_has_scoped,
-                ..Default::default()
-            },
-            vapor,
-            scope_id: Some(scope_id.clone()),
-        };
-
-        match sfc_compile(&descriptor, compile_opts) {
-            Ok(result) => {
-                let macro_artifacts = macro_artifacts_to_napi(result.macro_artifacts);
-                success_count.fetch_add(1, Ordering::Relaxed);
-                push_result(BatchFileResultNapi {
-                    path: filename.clone(),
-                    code: result.code.into(),
-                    css: result.css.map(Into::into),
-                    scope_id: scope_id.into(),
-                    has_scoped: actual_has_scoped,
-                    errors: result
-                        .errors
-                        .into_iter()
-                        .map(|e| e.message.into())
-                        .collect(),
-                    warnings: result
-                        .warnings
-                        .into_iter()
-                        .map(|e| e.message.into())
-                        .collect(),
-                    template_hash: template_hash.clone(),
-                    style_hash: style_hash.clone(),
-                    script_hash: script_hash.clone(),
-                    styles,
-                    custom_blocks,
-                    macro_artifacts,
-                });
-            }
-            Err(e) => {
-                failed_count.fetch_add(1, Ordering::Relaxed);
-                push_result(BatchFileResultNapi {
-                    path: filename.clone(),
-                    code: String::new(),
-                    css: None,
-                    scope_id: scope_id.into(),
-                    has_scoped: actual_has_scoped,
-                    errors: vec![e.message.into()],
-                    warnings: vec![],
-                    template_hash,
-                    style_hash,
-                    script_hash,
-                    styles,
-                    custom_blocks,
-                    macro_artifacts: vec![],
-                });
-            }
-        }
-    });
-
-    let elapsed = start.elapsed();
-    let final_results = match results.into_inner() {
-        Ok(results) => results,
-        Err(poisoned) => poisoned.into_inner(),
-    };
-
-    Ok(BatchCompileResultWithFilesNapi {
-        results: final_results,
-        success_count: success_count.load(Ordering::Relaxed) as u32,
-        failed_count: failed_count.load(Ordering::Relaxed) as u32,
-        time_ms: elapsed.as_secs_f64() * 1000.0,
-    })
-}
-
-/// CSS compile options for NAPI
-#[napi(object)]
-#[derive(Default)]
-pub struct CssCompileOptionsNapi {
-    /// Filename for error reporting
-    pub filename: Option<String>,
-    /// Whether to apply scoped CSS transformation
-    pub scoped: Option<bool>,
-    /// Scope ID for scoped CSS (e.g., "data-v-abc123"). Must be the full attribute name.
-    pub scope_id: Option<String>,
-    /// Whether to generate source maps
-    pub source_map: Option<bool>,
-    /// Whether to minify the output
-    pub minify: Option<bool>,
-    /// Whether to enable CSS Modules transforms
-    pub css_modules: Option<bool>,
-    /// Whether to enable custom media query resolution
-    pub custom_media: Option<bool>,
-    /// Browser targets for autoprefixing
-    pub targets: Option<CssTargetsNapi>,
-}
-
-/// Browser targets for CSS autoprefixing
-#[napi(object)]
-#[derive(Default)]
-pub struct CssTargetsNapi {
-    pub chrome: Option<u32>,
-    pub firefox: Option<u32>,
-    pub safari: Option<u32>,
-    pub edge: Option<u32>,
-    pub ios: Option<u32>,
-    pub android: Option<u32>,
-}
-
-/// CSS compile result for NAPI
-#[napi(object)]
-pub struct CssCompileResultNapi {
-    /// Compiled CSS code
-    pub code: String,
-    /// Source map (if requested)
-    pub map: Option<String>,
-    /// CSS variables found (from v-bind())
-    pub css_vars: Vec<String>,
-    /// Errors during compilation
-    pub errors: Vec<String>,
-    /// Warnings during compilation
-    pub warnings: Vec<String>,
-}
-
-/// Compile a CSS string with scoped CSS, v-bind() extraction, and optional minification.
-/// Unlike `compileSfc`, the `scopeId` is used as-is without stripping the "data-v-" prefix.
-/// Callers must pass the full attribute name (e.g., "data-v-abc123").
-#[napi(js_name = "compileCss")]
-pub fn compile_css_napi(
-    source: String,
-    options: Option<CssCompileOptionsNapi>,
-) -> Result<CssCompileResultNapi> {
-    use vize_atelier_sfc::{CssCompileOptions, CssTargets, compile_css};
-
-    let opts = options.unwrap_or_default();
-
-    let targets = opts.targets.map(|t| CssTargets {
-        chrome: t.chrome,
-        firefox: t.firefox,
-        safari: t.safari,
-        edge: t.edge,
-        ios: t.ios,
-        android: t.android,
-    });
-
-    let compile_opts = CssCompileOptions {
-        filename: opts.filename.map(Into::into),
-        scoped: opts.scoped.unwrap_or(false),
-        scope_id: opts.scope_id.map(Into::into),
-        source_map: opts.source_map.unwrap_or(false),
-        minify: opts.minify.unwrap_or(false),
-        css_modules: opts.css_modules.unwrap_or(false),
-        custom_media: opts.custom_media.unwrap_or(false),
-        targets,
-    };
-
-    let result = compile_css(&source, &compile_opts);
-
-    Ok(CssCompileResultNapi {
-        code: result.code.into(),
-        map: result.map.map(Into::into),
-        css_vars: result.css_vars.into_iter().map(Into::into).collect(),
-        errors: result.errors.into_iter().map(Into::into).collect(),
-        warnings: result.warnings.into_iter().map(Into::into).collect(),
-    })
-}
+mod batch;
+mod batch_results;
+mod bundler;
+mod compile;
+mod css;
+mod parse;
+mod types;
+
+pub use batch::*;
+pub use batch_results::*;
+pub use bundler::*;
+pub use compile::*;
+pub use css::*;
+pub use parse::*;
+pub use types::*;
